+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 - Log errors if no error handler is registered
 - Log warnings if an unsupported server management protocol version is detected
 
+## Relevant Minecraft changes.
+- Gamerules are now stored in registries. This means that all gamerule names have been changed. You can find the new
+names [here](https://www.minecraft.net/en-us/article/minecraft-snapshot-25w44a). `server.hasGameRulesRegistry()` can be 
+used to determine whether the server uses the old or the new names.
+
 # 3.1.0
 ## Improvements
 - Add `SERVER_ACTIVITY` notification